backend/requirements.txt
django==5.1
django-rest-framework==0.1.0
django-cors-headers==4.4.0
drf-writable-nested==0.7.0
psycopg2-binary==2.9.9
cpf-cnpj-validate==1.4
faker==27.0.0
pytest-django==4.8.0
django-extensions==3.2.3